using corpora in usermask for #2084
